requirements.txt
pyserial>=3.5
requests>=2.28.0
pyyaml>=6.0
geopandas>=0.14.0
matplotlib>=3.7.0
shapely>=2.0.0
fiona>=1.9.0
geodatasets>=0.3.0
